Created a ShootTask for the Ghost Entity
Created a ShootTask which is called in ChaseTask. When the ghost is in a certain range of the target, it chases (as per the original game engine), and also calls the ShootTask which prints a message to the terminal. The chasing functionality is yet to be removed, and development will need to be done on the ShootTask after consultation with the Projectiles team.